Refactor HubConnectionHandler logging
