+++ src/SignalR/server/Core/src/HubConnectionHandler.cs
@@ -14,6 +14,7 @@ using Microsoft.AspNetCore.SignalR.Protocol;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Log = Microsoft.AspNetCore.SignalR.HubConnectionHandlerLog;
 
 namespace Microsoft.AspNetCore.SignalR
 {
@@ -333,57 +334,6 @@ namespace Microsoft.AspNetCore.SignalR
                     // before yielding the read again.
                     input.AdvanceTo(buffer.Start, buffer.End);
                 }
-            }
-        }
-
-        private static class Log
-        {
-            private static readonly Action<ILogger, string, Exception?> _errorDispatchingHubEvent =
-                LoggerMessage.Define<string>(LogLevel.Error, new EventId(1, "ErrorDispatchingHubEvent"), "Error when dispatching '{HubMethod}' on hub.");
-
-            private static readonly Action<ILogger, Exception?> _errorProcessingRequest =
-                LoggerMessage.Define(LogLevel.Debug, new EventId(2, "ErrorProcessingRequest"), "Error when processing requests.");
-
-            private static readonly Action<ILogger, Exception?> _abortFailed =
-                LoggerMessage.Define(LogLevel.Trace, new EventId(3, "AbortFailed"), "Abort callback failed.");
-
-            private static readonly Action<ILogger, Exception?> _errorSendingClose =
-                LoggerMessage.Define(LogLevel.Debug, new EventId(4, "ErrorSendingClose"), "Error when sending Close message.");
-
-            private static readonly Action<ILogger, Exception?> _connectedStarting =
-                LoggerMessage.Define(LogLevel.Debug, new EventId(5, "ConnectedStarting"), "OnConnectedAsync started.");
-
-            private static readonly Action<ILogger, Exception?> _connectedEnding =
-                LoggerMessage.Define(LogLevel.Debug, new EventId(6, "ConnectedEnding"), "OnConnectedAsync ending.");
-
-            public static void ErrorDispatchingHubEvent(ILogger logger, string hubMethod, Exception exception)
-            {
-                _errorDispatchingHubEvent(logger, hubMethod, exception);
-            }
-
-            public static void ErrorProcessingRequest(ILogger logger, Exception exception)
-            {
-                _errorProcessingRequest(logger, exception);
-            }
-
-            public static void AbortFailed(ILogger logger, Exception exception)
-            {
-                _abortFailed(logger, exception);
-            }
-
-            public static void ErrorSendingClose(ILogger logger, Exception exception)
-            {
-                _errorSendingClose(logger, exception);
-            }
-
-            public static void ConnectedStarting(ILogger logger)
-            {
-                _connectedStarting(logger, null);
-            }
-
-            public static void ConnectedEnding(ILogger logger)
-            {
-                _connectedEnding(logger, null);
             }
         }
     }
